backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pandas==2.2.0
scikit-learn==1.4.0
python-socketio==5.11.1
python-multipart==0.0.9
joblib==1.3.2
python-dotenv==1.0.1
redis==5.0.1